+++ www/go/modules/community/addressbook/views/extjs3/SystemSettingsPanel.js
@@ -25,12 +25,14 @@ go.modules.community.addressbook.SystemSettingsPanel = Ext.extend(go.systemsetti
 					name: "autoLinkEmail",
 					disabled: !GO.savemailas,
 					hint: t("Warning: this will copy e-mails to the Group-Office storage and will therefore increase disk space usage. The e-mail will be visible to all people that can view the contact too.")
-				},{
-					hideLabel: true,
-					xtype: "checkbox",
-					boxLabel: t("Restrict export to administrators"),
-					name: "restrictExportToAdmins"
 				}]
+			// As per 6.6, this permission has been dropped in favor of the module permission mayExportContacts
+				// ,{
+				// 	hideLabel: true,
+				// 	xtype: "checkbox",
+				// 	boxLabel: t("Restrict export to administrators"),
+				// 	name: "restrictExportToAdmins"
+				// }]
 		}];
 
 		go.modules.community.addressbook.SystemSettingsPanel.superclass.initComponent.call(this);